backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pywebview==5.3.2
requests>=2.31.0
